+++ lib/canary/plugs.ex
@@ -67,8 +67,8 @@ defmodule Canary.Plugs do
   * `:preload` - Specifies association(s) to preload
   * `:id_name` - Specifies the name of the id in `conn.params`, defaults to "id"
   * `:id_field` - Specifies the name of the ID field in the database for searching :id_name value, defaults to "id".
-  * `:persisted` - Specifies the resource should always be loaded from the database, defaults to false
-  * `:required` - Same as `:persisted` but with not found handler - even for :index, :new or :create action
+  * `:persisted` - Deprecated, specifies the resource should always be loaded from the database, defaults to false
+  * `:required` - Specifies if the resource is required, when it's not found it will call the not found handler - even for :index, :new or :create action, defaults to true
   * `:not_found_handler` - Specify a handler function to be called if the resource is not found
 
 
@@ -89,6 +89,7 @@ defmodule Canary.Plugs do
   """
   @spec load_resource(Plug.Conn.t(), Plug.opts()) :: Plug.Conn.t()
   def load_resource(conn, opts) do
+    validate_opts(opts)
     action = get_action(conn)
 
     if action_valid?(action, opts) do
@@ -119,7 +120,7 @@ defmodule Canary.Plugs do
           fetch_resource(conn, opts)
       end
 
-    Plug.Conn.assign(conn, get_resource_name(conn, opts), loaded_resource)
+    Plug.Conn.assign(conn, get_resource_name(action, opts), loaded_resource)
   end
 
   @doc """
@@ -161,6 +162,7 @@ defmodule Canary.Plugs do
   """
   @spec authorize_controller(Plug.Conn.t(), Plug.opts()) :: Plug.Conn.t()
   def authorize_controller(conn, opts) do
+    validate_opts(opts)
     action = get_action(conn)
 
     if action_valid?(action, opts) do
@@ -258,6 +260,7 @@ defmodule Canary.Plugs do
   """
   @spec authorize_resource(Plug.Conn.t(), Plug.opts()) :: Plug.Conn.t()
   def authorize_resource(conn, opts) do
+    validate_opts(opts)
     action = get_action(conn)
 
     if action_valid?(action, opts) do
@@ -275,19 +278,12 @@ defmodule Canary.Plugs do
     action = get_action(conn)
     is_persisted = persisted?(opts)
 
-    non_id_actions =
-      if opts[:non_id_actions] do
-        Enum.concat([:index, :new, :create], opts[:non_id_actions])
-      else
-        [:index, :new, :create]
-      end
-
     resource =
       cond do
         is_persisted ->
           fetch_resource(conn, opts)
 
-        action in non_id_actions ->
+        action in non_id_actions(opts) ->
           opts[:model]
 
         true ->
@@ -342,6 +338,7 @@ defmodule Canary.Plugs do
   ```
   """
   def load_and_authorize_resource(conn, opts) do
+    validate_opts(opts)
     action = get_action(conn)
 
     if action_valid?(action, opts) do
@@ -353,12 +350,10 @@ defmodule Canary.Plugs do
 
   defp do_load_and_authorize_resource(conn, opts) do
     conn
-    # skip not_found_handler so auth handler can catch first if needed
-    |> Map.put(:skip_canary_handler, true)
-    |> load_resource(opts)
-    # allow auth handling
-    |> Map.delete(:skip_canary_handler)
-    |> authorize_resource(opts)
+    # not found is handled after authorization so the unauthorized handler can catch first
+    |> do_load_resource(opts)
+    |> do_authorize_resource(opts)
+    |> handle_unauthorized(opts)
     |> maybe_handle_not_found(opts)
     |> purge_resource_if_unauthorized(opts)
   end
@@ -371,7 +366,7 @@ defmodule Canary.Plugs do
     do: conn
 
   defp purge_resource_if_unauthorized(%{assigns: %{authorized: false}} = conn, opts),
-    do: Plug.Conn.assign(conn, get_resource_name(conn, opts), nil)
+    do: Plug.Conn.assign(conn, get_resource_name(get_action(conn), opts), nil)
 
   defp fetch_resource(conn, opts) do
     repo = Application.get_env(:canary, :repo)
@@ -380,7 +375,7 @@ defmodule Canary.Plugs do
 
     get_map_args = %{String.to_atom(field_name) => get_resource_id(conn, opts)}
 
-    case Map.fetch(conn.assigns, get_resource_name(conn, opts)) do
+    case Map.fetch(conn.assigns, get_resource_name(get_action(conn), opts)) do
       :error ->
         repo.get_by(opts[:model], get_map_args)
         |> preload_if_needed(repo, opts)
@@ -404,7 +399,7 @@ defmodule Canary.Plugs do
   defp fetch_all(conn, opts) do
     repo = Application.get_env(:canary, :repo)
 
-    resource_name = get_resource_name(conn, opts)
+    resource_name = get_resource_name(get_action(conn), opts)
 
     # check if a resource is already loaded at the key
     case Map.fetch(conn.assigns, resource_name) do
@@ -427,57 +422,14 @@ defmodule Canary.Plugs do
     end
   end
 
-  defp persisted?(opts) do
-    !!Keyword.get(opts, :persisted, false) || !!Keyword.get(opts, :required, false)
-  end
-
-  defp get_resource_name(conn, opts) do
-    case opts[:as] do
-      nil ->
-        opts[:model]
-        |> Module.split()
-        |> List.last()
-        |> Macro.underscore()
-        |> pluralize_if_needed(conn, opts)
-        |> String.to_atom()
-
-      as ->
-        as
-    end
-  end
-
-  defp pluralize_if_needed(name, conn, opts) do
-    if get_action(conn) in [:index] and not persisted?(opts) do
-      name <> "s"
-    else
-      name
-    end
-  end
-
   defp handle_unauthorized(%{assigns: %{authorized: true}} = conn, _opts),
     do: conn
 
   defp handle_unauthorized(%{assigns: %{authorized: false}} = conn, opts),
     do: apply_error_handler(conn, :unauthorized_handler, opts)
 
-  defp handle_not_found(%{skip_canary_handler: true} = conn, _opts) do
-    conn
-  end
-
   defp handle_not_found(conn, opts) do
-    action = get_action(conn)
-
-    non_id_actions =
-      if opts[:non_id_actions] do
-        Enum.concat([:index, :new, :create], opts[:non_id_actions])
-      else
-        [:index, :new, :create]
-      end
-
-    is_required = required?(opts)
-    resource_name = Map.get(conn.assigns, get_resource_name(conn, opts))
-
-    if is_nil(resource_name) and (is_required or action not in non_id_actions) do
+    if apply_handle_not_found?(get_action(conn), conn.assigns, opts) do
       apply_error_handler(conn, :not_found_handler, opts)
     else
       conn